Refactor buffer.c and new struct for parser.c
I've refactored almost all of the basic buffer functions (which are
in buffer.c) to accept a pointer to a buffer for more unification.
And also one of the imported changes - added a structure for tokens
in parser.c and parser.h to make this part of a program more
structured.