Fix 1 bug and compilation error
first - a first bug is related to not freeing the allocated memory
for arguments in time, that is, the memory was freed after some time,
and not immediately after use, as a result of which the p function
did not want to display the contents of the buffer.
second - the compilation error is related to a typo in the parser.c
file when the structure is iniitalized

both issues fixed.